IMA loop as worker added
