+++ agent/cli.js
@@ -336,7 +336,8 @@ function parse( joExternalHandlers, argv ) {
             console.log( soi + cc.debug( "--" ) + cc.bright( "s2s-transfer" ) + cc.debug( ".................." ) + cc.notice( "Do single " ) + cc.attention( "money transfer loop" ) + cc.notice( " from " ) + cc.note( "S-chain" ) + cc.notice( " to " ) + cc.note( "S-chain" ) + cc.notice( "." ) );
             console.log( soi + cc.debug( "--" ) + cc.bright( "with-metadata" ) + cc.debug( "................." ) + cc.notice( "Makes " ) + cc.bright( "ERC721" ) + cc.notice( " transfer using special version of " ) + cc.bright( "Token Manager" ) + cc.notice( " to transfer token metadata." ) );
             console.log( soi + cc.debug( "--" ) + cc.bright( "transfer" ) + cc.debug( "......................" ) + cc.notice( "Run single " ) + cc.note( "M<->S" ) + cc.notice( " and, optionally, " ) + cc.note( "S->S" ) + cc.notice( " transfer loop iteration" ) + cc.notice( "." ) );
-            console.log( soi + cc.debug( "--" ) + cc.bright( "loop" ) + cc.debug( ".........................." ) + cc.notice( "Run " ) + cc.note( "M<->S" ) + cc.notice( " and, optionally, " ) + cc.note( "S->S" ) + cc.notice( " transfer loop." ) );
+            console.log( soi + cc.debug( "--" ) + cc.bright( "loop" ) + cc.debug( ".........................." ) + cc.notice( "Run " ) + cc.note( "M<->S" ) + cc.notice( " and, optionally, " ) + cc.note( "S->S" ) + cc.notice( " transfer loops in parallel threads." ) );
+            console.log( soi + cc.debug( "--" ) + cc.bright( "simple-loop" ) + cc.debug( "..................." ) + cc.notice( "Run " ) + cc.note( "M<->S" ) + cc.notice( " and, optionally, " ) + cc.note( "S->S" ) + cc.notice( " transfer loops in main thread only." ) );
             console.log( soi + cc.debug( "--" ) + cc.bright( "load-node-config" ) + cc.sunny( "=" ) + cc.success( "path" ) + cc.debug( "........." ) + cc.notice( "Use specified " ) + cc.note( "S-Chain" ) + cc.notice( " node JSON configuration file to load parameters" ) + cc.debug( "(like " ) + cc.attention( "node index" ) + cc.debug( ", " ) + cc.attention( "nodes count" ) + cc.debug( ")" ) + cc.notice( "." ) );
             //
             console.log( cc.sunny( "ADDITIONAL ACTION" ) + cc.info( " options:" ) );
@@ -1253,6 +1254,7 @@ function parse( joExternalHandlers, argv ) {
             joArg.name == "s2s-transfer" ||
             joArg.name == "transfer" ||
             joArg.name == "loop" ||
+            joArg.name == "simple-loop" ||
             joArg.name == "browse-s-chain" ||
             joArg.name == "browse-skale-network" ||
             joArg.name == "browse-connected-schains" ||
@@ -1516,109 +1518,6 @@ function ima_common_init() {
     // token_manager_linker_address               --> token_manager_linker_abi
     // message_proxy_mainnet_address              --> message_proxy_mainnet_abi
     // message_proxy_chain_address                --> message_proxy_chain_abi
-
-    //
-    if( imaState.chainProperties.mn.strURL && typeof imaState.chainProperties.mn.strURL == "string" && imaState.chainProperties.mn.strURL.length > 0 ) {
-        const u = imaState.chainProperties.mn.strURL;
-        async_check_url_at_startup( u, "Main-net" );
-        imaState.chainProperties.mn.w3 = getWeb3FromURL( u, log );
-    } else {
-        log.write(
-            cc.error( "WARNING:" ) + cc.warning( " No " ) + cc.note( "Main-net" ) +
-            cc.warning( " URL specified in command line arguments" ) +
-            cc.debug( "(needed for particular operations only)" ) +
-            "\n" );
-    }
-    //
-    if( imaState.chainProperties.sc.strURL && typeof imaState.chainProperties.sc.strURL == "string" && imaState.chainProperties.sc.strURL.length > 0 ) {
-        const u = imaState.chainProperties.sc.strURL;
-        async_check_url_at_startup( u, "S-Chain" );
-        imaState.chainProperties.sc.w3 = getWeb3FromURL( u, log );
-    } else {
-        log.write(
-            cc.error( "WARNING:" ) + cc.warning( " No " ) + cc.note( "S-Chain" ) +
-            cc.warning( " URL specified in command line arguments" ) +
-            cc.debug( "(needed for particular operations only)" ) +
-            "\n" );
-    }
-    //
-    if( imaState.chainProperties.tc.strURL && typeof imaState.chainProperties.tc.strURL == "string" && imaState.chainProperties.tc.strURL.length > 0 ) {
-        const u = imaState.chainProperties.tc.strURL;
-        async_check_url_at_startup( u, "S<->S Target S-Chain" );
-        imaState.chainProperties.tc.w3 = getWeb3FromURL( u, log );
-    } else {
-        log.write(
-            cc.error( "WARNING:" ) + cc.warning( " No " ) + cc.note( "S<->S Target S-Chain" ) +
-            cc.warning( " URL specified in command line arguments" ) +
-            cc.debug( "(needed for particular operations only)" ) +
-            "\n" );
-    }
-    //
-    if( imaState.chainProperties.mn.bHaveAbiIMA ) {
-        imaState.jo_deposit_box_eth = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_eth_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_eth_address ); // only main net
-        imaState.jo_deposit_box_erc20 = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc20_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc20_address ); // only main net
-        imaState.jo_deposit_box_erc721 = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_address ); // only main net
-        imaState.jo_deposit_box_erc1155 = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc1155_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc1155_address ); // only main net
-        imaState.jo_deposit_box_erc721_with_metadata = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_with_metadata_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_with_metadata_address ); // only main net
-        imaState.jo_community_pool = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.community_pool_abi, imaState.chainProperties.mn.joAbiIMA.community_pool_address ); // only main net
-        imaState.jo_linker = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.linker_abi, imaState.chainProperties.mn.joAbiIMA.linker_address ); // only main net
-        imaState.jo_message_proxy_main_net = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.message_proxy_mainnet_abi, imaState.chainProperties.mn.joAbiIMA.message_proxy_mainnet_address );
-    }
-    if( imaState.chainProperties.sc.bHaveAbiIMA ) {
-        imaState.jo_token_manager_eth = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_eth_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_eth_address ); // only s-chain
-        imaState.jo_token_manager_erc20 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc20_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc20_address ); // only s-chain
-        imaState.jo_token_manager_erc721 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_address ); // only s-chain
-        imaState.jo_token_manager_erc1155 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc1155_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc1155_address ); // only s-chain
-        imaState.jo_token_manager_erc721_with_metadata = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_with_metadata_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_with_metadata_address ); // only s-chain
-        imaState.jo_community_locker = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.community_locker_abi, imaState.chainProperties.sc.joAbiIMA.community_locker_address ); // only s-chain
-        imaState.jo_message_proxy_s_chain = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.message_proxy_chain_abi, imaState.chainProperties.sc.joAbiIMA.message_proxy_chain_address );
-        imaState.jo_token_manager_linker = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_linker_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_linker_address );
-        imaState.eth_erc20 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.eth_erc20_abi, imaState.chainProperties.sc.joAbiIMA.eth_erc20_address ); // only s-chain
-        // imaState.eth_erc721 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.eth_erc721_abi, imaState.chainProperties.sc.joAbiIMA.eth_erc721_address ); // only s-chain
-        // imaState.eth_erc1155 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.eth_erc1155_abi, imaState.chainProperties.sc.joAbiIMA.eth_erc721_address ); // only s-chain
-    }
-    if( imaState.chainProperties.tc.bHaveAbiIMA ) {
-        // imaState.jo_token_manager_eth_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_eth_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_eth_address ); // only s-chain
-        imaState.jo_token_manager_erc20_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc20_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc20_address ); // only s-chain
-        imaState.jo_token_manager_erc721_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_address ); // only s-chain
-        imaState.jo_token_manager_erc1155_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc1155_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc1155_address ); // only s-chain
-        imaState.jo_token_manager_erc721_with_metadata_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_with_metadata_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_with_metadata_address ); // only s-chain
-        imaState.jo_community_locker_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.community_locker_abi, imaState.chainProperties.tc.joAbiIMA.community_locker_address ); // only s-chain
-        imaState.jo_message_proxy_s_chain_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.message_proxy_chain_abi, imaState.chainProperties.tc.joAbiIMA.message_proxy_chain_address );
-        imaState.jo_token_manager_linker_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_linker_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_linker_address );
-        imaState.eth_erc20_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.eth_erc20_abi, imaState.chainProperties.tc.joAbiIMA.eth_erc20_address ); // only s-chain
-        // imaState.eth_erc721_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.eth_erc721_abi, imaState.chainProperties.tc.joAbiIMA.eth_erc721_address ); // only s-chain
-        // imaState.eth_erc1155_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.eth_erc1155_abi, imaState.chainProperties.tc.joAbiIMA.eth_erc721_address ); // only s-chain
-    }
-    if( imaState.bHaveSkaleManagerABI ) {
-        imaState.jo_constants_holder = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.constants_holder_abi, imaState.joAbiSkaleManager.constants_holder_address );
-        // jo_contract_manager
-        // jo_decryption
-        // jo_delegation_controller
-        // jo_delegation_period_manager
-        // jo_distributor
-        // jo_ecdh
-        // jo_manager_data
-        // jo_monitors_functionality
-        imaState.jo_nodes = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.nodes_abi, imaState.joAbiSkaleManager.nodes_address );
-        // jo_pricing
-        // jo_punisher
-        imaState.jo_key_storage = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.key_storage_abi, imaState.joAbiSkaleManager.key_storage_address );
-        imaState.jo_schains = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.schains_abi, imaState.joAbiSkaleManager.schains_address );
-        imaState.jo_schains_internal = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.schains_internal_abi, imaState.joAbiSkaleManager.schains_internal_address );
-        // jo_schains_functionality
-        // jo_schains_functionality_internal
-        imaState.jo_skale_dkg = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.skale_d_k_g_abi, imaState.joAbiSkaleManager.skale_d_k_g_address );
-        imaState.jo_skale_manager = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.skale_manager_abi, imaState.joAbiSkaleManager.skale_manager_address );
-        imaState.jo_skale_token = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.skale_token_abi, imaState.joAbiSkaleManager.skale_token_address );
-        // jo_skale_verifier
-        // jo_slashing_table
-        // jo_time_helpers
-        // jo_time_helpers_with_debug
-        // jo_token_state
-        imaState.jo_validator_service = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.validator_service_abi, imaState.joAbiSkaleManager.validator_service_address );
-        imaState.jo_wallets = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.wallets_abi, imaState.joAbiSkaleManager.wallets_address );
-    } // if( imaState.bHaveSkaleManagerABI )
 
     const oct = function( joContract ) { // optional contract address
         if( joContract && "options" in joContract && "address" in joContract.options )
@@ -2302,10 +2201,126 @@ function ima_common_init() {
         log.write( cc.info( "IMA JSON RPC server port is" ) + cc.debug( "...,,,,,,,,,,,............" ) + ( ( imaState.nJsonRpcPort > 0 ) ? cc.info( imaState.nJsonRpcPort ) : cc.error( "disabled" ) ) + "\n" );
         log.write( cc.info( "Cross-IMA mode is" ) + cc.debug( "...................................." ) + ( imaState.isCrossImaBlsMode ? cc.success( "enabled" ) : cc.error( "disabled" ) ) + "\n" );
     }
-    //
-    //
-    //
 } // ima_common_init
+
+function ima_w3_init() {
+    if( ( !imaState.chainProperties.mn.w3 ) &&
+        imaState.chainProperties.mn.strURL &&
+        typeof imaState.chainProperties.mn.strURL == "string" &&
+        imaState.chainProperties.mn.strURL.length > 0
+    ) {
+        const u = imaState.chainProperties.mn.strURL;
+        async_check_url_at_startup( u, "Main-net" );
+        imaState.chainProperties.mn.w3 = getWeb3FromURL( u, log );
+    } else {
+        log.write(
+            cc.error( "WARNING:" ) + cc.warning( " No " ) + cc.note( "Main-net" ) +
+            cc.warning( " URL specified in command line arguments" ) +
+            cc.debug( "(needed for particular operations only)" ) +
+            "\n" );
+    }
+    //
+    if( ( !imaState.chainProperties.sc.w3 ) &&
+        imaState.chainProperties.sc.strURL &&
+        typeof imaState.chainProperties.sc.strURL == "string" &&
+        imaState.chainProperties.sc.strURL.length > 0
+    ) {
+        const u = imaState.chainProperties.sc.strURL;
+        async_check_url_at_startup( u, "S-Chain" );
+        imaState.chainProperties.sc.w3 = getWeb3FromURL( u, log );
+    } else {
+        log.write(
+            cc.error( "WARNING:" ) + cc.warning( " No " ) + cc.note( "S-Chain" ) +
+            cc.warning( " URL specified in command line arguments" ) +
+            cc.debug( "(needed for particular operations only)" ) +
+            "\n" );
+    }
+    //
+    if( ( !imaState.chainProperties.tc.w3 ) &&
+        imaState.chainProperties.tc.strURL &&
+        typeof imaState.chainProperties.tc.strURL == "string" &&
+        imaState.chainProperties.tc.strURL.length > 0
+    ) {
+        const u = imaState.chainProperties.tc.strURL;
+        async_check_url_at_startup( u, "S<->S Target S-Chain" );
+        imaState.chainProperties.tc.w3 = getWeb3FromURL( u, log );
+    } else {
+        log.write(
+            cc.error( "WARNING:" ) + cc.warning( " No " ) + cc.note( "S<->S Target S-Chain" ) +
+            cc.warning( " URL specified in command line arguments" ) +
+            cc.debug( "(needed for particular operations only)" ) +
+            "\n" );
+    }
+} // ima_w3_init
+
+function ima_contracts_init() {
+    ima_w3_init();
+    if( imaState.chainProperties.mn.bHaveAbiIMA && ( !imaState.jo_deposit_box_eth ) ) {
+        imaState.jo_deposit_box_eth = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_eth_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_eth_address ); // only main net
+        imaState.jo_deposit_box_erc20 = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc20_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc20_address ); // only main net
+        imaState.jo_deposit_box_erc721 = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_address ); // only main net
+        imaState.jo_deposit_box_erc1155 = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc1155_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc1155_address ); // only main net
+        imaState.jo_deposit_box_erc721_with_metadata = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_with_metadata_abi, imaState.chainProperties.mn.joAbiIMA.deposit_box_erc721_with_metadata_address ); // only main net
+        imaState.jo_community_pool = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.community_pool_abi, imaState.chainProperties.mn.joAbiIMA.community_pool_address ); // only main net
+        imaState.jo_linker = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.linker_abi, imaState.chainProperties.mn.joAbiIMA.linker_address ); // only main net
+        imaState.jo_message_proxy_main_net = new imaState.chainProperties.mn.w3.eth.Contract( imaState.chainProperties.mn.joAbiIMA.message_proxy_mainnet_abi, imaState.chainProperties.mn.joAbiIMA.message_proxy_mainnet_address );
+    }
+    if( imaState.chainProperties.sc.bHaveAbiIMA && ( !imaState.jo_token_manager_eth ) ) {
+        imaState.jo_token_manager_eth = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_eth_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_eth_address ); // only s-chain
+        imaState.jo_token_manager_erc20 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc20_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc20_address ); // only s-chain
+        imaState.jo_token_manager_erc721 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_address ); // only s-chain
+        imaState.jo_token_manager_erc1155 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc1155_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc1155_address ); // only s-chain
+        imaState.jo_token_manager_erc721_with_metadata = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_with_metadata_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_erc721_with_metadata_address ); // only s-chain
+        imaState.jo_community_locker = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.community_locker_abi, imaState.chainProperties.sc.joAbiIMA.community_locker_address ); // only s-chain
+        imaState.jo_message_proxy_s_chain = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.message_proxy_chain_abi, imaState.chainProperties.sc.joAbiIMA.message_proxy_chain_address );
+        imaState.jo_token_manager_linker = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.token_manager_linker_abi, imaState.chainProperties.sc.joAbiIMA.token_manager_linker_address );
+        imaState.eth_erc20 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.eth_erc20_abi, imaState.chainProperties.sc.joAbiIMA.eth_erc20_address ); // only s-chain
+        // imaState.eth_erc721 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.eth_erc721_abi, imaState.chainProperties.sc.joAbiIMA.eth_erc721_address ); // only s-chain
+        // imaState.eth_erc1155 = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.sc.joAbiIMA.eth_erc1155_abi, imaState.chainProperties.sc.joAbiIMA.eth_erc721_address ); // only s-chain
+    }
+    if( imaState.chainProperties.tc.bHaveAbiIMA && ( !imaState.jo_token_manager_erc20_target ) ) {
+        // imaState.jo_token_manager_eth_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_eth_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_eth_address ); // only s-chain
+        imaState.jo_token_manager_erc20_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc20_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc20_address ); // only s-chain
+        imaState.jo_token_manager_erc721_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_address ); // only s-chain
+        imaState.jo_token_manager_erc1155_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc1155_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc1155_address ); // only s-chain
+        imaState.jo_token_manager_erc721_with_metadata_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_with_metadata_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_erc721_with_metadata_address ); // only s-chain
+        imaState.jo_community_locker_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.community_locker_abi, imaState.chainProperties.tc.joAbiIMA.community_locker_address ); // only s-chain
+        imaState.jo_message_proxy_s_chain_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.message_proxy_chain_abi, imaState.chainProperties.tc.joAbiIMA.message_proxy_chain_address );
+        imaState.jo_token_manager_linker_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.token_manager_linker_abi, imaState.chainProperties.tc.joAbiIMA.token_manager_linker_address );
+        imaState.eth_erc20_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.eth_erc20_abi, imaState.chainProperties.tc.joAbiIMA.eth_erc20_address ); // only s-chain
+        // imaState.eth_erc721_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.eth_erc721_abi, imaState.chainProperties.tc.joAbiIMA.eth_erc721_address ); // only s-chain
+        // imaState.eth_erc1155_target = new imaState.chainProperties.sc.w3.eth.Contract( imaState.chainProperties.tc.joAbiIMA.eth_erc1155_abi, imaState.chainProperties.tc.joAbiIMA.eth_erc721_address ); // only s-chain
+    }
+    if( imaState.bHaveSkaleManagerABI && ( !imaState.jo_constants_holder ) ) {
+        imaState.jo_constants_holder = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.constants_holder_abi, imaState.joAbiSkaleManager.constants_holder_address );
+        // jo_contract_manager
+        // jo_decryption
+        // jo_delegation_controller
+        // jo_delegation_period_manager
+        // jo_distributor
+        // jo_ecdh
+        // jo_manager_data
+        // jo_monitors_functionality
+        imaState.jo_nodes = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.nodes_abi, imaState.joAbiSkaleManager.nodes_address );
+        // jo_pricing
+        // jo_punisher
+        imaState.jo_key_storage = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.key_storage_abi, imaState.joAbiSkaleManager.key_storage_address );
+        imaState.jo_schains = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.schains_abi, imaState.joAbiSkaleManager.schains_address );
+        imaState.jo_schains_internal = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.schains_internal_abi, imaState.joAbiSkaleManager.schains_internal_address );
+        // jo_schains_functionality
+        // jo_schains_functionality_internal
+        imaState.jo_skale_dkg = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.skale_d_k_g_abi, imaState.joAbiSkaleManager.skale_d_k_g_address );
+        imaState.jo_skale_manager = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.skale_manager_abi, imaState.joAbiSkaleManager.skale_manager_address );
+        imaState.jo_skale_token = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.skale_token_abi, imaState.joAbiSkaleManager.skale_token_address );
+        // jo_skale_verifier
+        // jo_slashing_table
+        // jo_time_helpers
+        // jo_time_helpers_with_debug
+        // jo_token_state
+        imaState.jo_validator_service = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.validator_service_abi, imaState.joAbiSkaleManager.validator_service_address );
+        imaState.jo_wallets = new imaState.chainProperties.mn.w3.eth.Contract( imaState.joAbiSkaleManager.wallets_abi, imaState.joAbiSkaleManager.wallets_address );
+    } // if( imaState.bHaveSkaleManagerABI )
+} // ima_contracts_init
 
 module.exports = {
     init: init,
@@ -2317,5 +2332,7 @@ module.exports = {
     load_node_config: load_node_config,
     parse: parse,
     ima_common_init: ima_common_init,
+    ima_w3_init: ima_w3_init,
+    ima_contracts_init: ima_contracts_init,
     getWeb3FromURL: getWeb3FromURL
 }; // module.exports
